fixed mpeg-ts http2 incorrect parsing

+++ core/m3u8lib/parser.py
@@ -4,7 +4,7 @@ import requests
 import os
 
 
-def fetch_playlist_links(session: requests.Session, playlist_url: str) -> List[str]:
+def fetch_playlist_links(session: requests.Session, playlist_url: str, keep: bool = False) -> List[str]:
     """ Fetches the m3u8 playlist from the playlist_url
 
     Parameters
@@ -29,7 +29,8 @@ def fetch_playlist_links(session: requests.Session, playlist_url: str) -> List[s
         temp = file.readlines()
         temp = [link.strip() for link in temp]
 
-    os.unlink("links.txt")
+    if not keep:
+        os.unlink("links.txt")
 
     parsed_url = urlparse(playlist_url)
     base_url: str = f"{parsed_url.scheme}://{parsed_url.netloc}{'/'.join(parsed_url.path.split('/')[:-1])}/"
